# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/meetings": {
      "get": {
        "tags": [
          "meetings"
        ],
        "summary": "List Meetings",
        "operationId": "list_meetings_meetings_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingListPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/meetings/recordings/cleanup/preview": {
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Preview Recording Cleanup",
        "operationId": "preview_recording_cleanup_meetings_recordings_cleanup_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordingCleanupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingCleanupPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/meetings/recordings/cleanup": {
      "post": {
        "tags": [
          "meetings"
        ],
        "summary": "Execute Recording Cleanup",
        "operationId": "execute_recording_cleanup_meetings_recordings_cleanup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordingCleanupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingCleanupExecuteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/meetings/{meeting_id}": {
      "get": {
        "tags": [
          "meetings"
        ],
        "summary": "Get Meeting",
        "operationId": "get_meeting_meetings__meeting_id__get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "meetings"
        ],
        "summary": "Update Meeting Title",
        "operationId": "update_meeting_title_meetings__meeting_id__patch",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meeting Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTitleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateMeetingTitleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "meetings"
        ],
        "summary": "Delete Meeting",
        "operationId": "delete_meeting_meetings__meeting_id__delete",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteMeetingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/meetings/{meeting_id}/recordings": {
      "get": {
        "tags": [
          "meetings"
        ],
        "summary": "List Recordings",
        "operationId": "list_recordings_meetings__meeting_id__recordings_get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecordingAssetItem"
                  },
                  "title": "Response List Recordings Meetings  Meeting Id  Recordings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/meetings/{meeting_id}/recordings/{role}": {
      "get": {
        "tags": [
          "meetings"
        ],
        "summary": "Serve Recording",
        "operationId": "serve_recording_meetings__meeting_id__recordings__role__get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meeting Id"
            }
          },
          {
            "name": "role",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "other",
                "self"
              ],
              "type": "string",
              "title": "Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DeleteMeetingResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warning"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "DeleteMeetingResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MeetingDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "ai_note": {
            "type": "string",
            "title": "Ai Note",
            "default": ""
          },
          "minutes": {
            "type": "string",
            "title": "Minutes",
            "default": ""
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "turns": {
            "items": {
              "$ref": "#/components/schemas/TurnItem"
            },
            "type": "array",
            "title": "Turns"
          },
          "reply_suggestions": {
            "items": {
              "$ref": "#/components/schemas/ReplySuggestionItem"
            },
            "type": "array",
            "title": "Reply Suggestions"
          },
          "recording_assets": {
            "items": {
              "$ref": "#/components/schemas/RecordingAssetItem"
            },
            "type": "array",
            "title": "Recording Assets"
          }
        },
        "type": "object",
        "required": [
          "id",
          "started_at",
          "status"
        ],
        "title": "MeetingDetail"
      },
      "MeetingListItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "has_ai_note": {
            "type": "boolean",
            "title": "Has Ai Note"
          },
          "has_recording": {
            "type": "boolean",
            "title": "Has Recording"
          }
        },
        "type": "object",
        "required": [
          "id",
          "started_at",
          "status",
          "has_ai_note",
          "has_recording"
        ],
        "title": "MeetingListItem",
        "description": "Lightweight meeting list item with asset indicator flags."
      },
      "MeetingListPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MeetingListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "MeetingListPage",
        "description": "Paginated meeting list response."
      },
...
      "RecordingAssetItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "enum": [
              "other",
              "self"
            ],
            "title": "Role"
          },
          "format": {
            "type": "string",
            "title": "Format",
            "default": "wav"
          },
          "sample_rate": {
            "type": "integer",
            "title": "Sample Rate",
            "default": 16000
          },
          "channels": {
            "type": "integer",
            "title": "Channels",
            "default": 1
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "role",
          "started_at"
        ],
        "title": "RecordingAssetItem"
      },
      "RecordingCleanupExecuteResponse": {
        "properties": {
          "candidate_meeting_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Candidate Meeting Ids"
          },
          "delete_count": {
            "type": "integer",
            "title": "Delete Count"
          },
          "delete_recording_bytes": {
            "type": "integer",
            "title": "Delete Recording Bytes"
          },
          "total_recording_bytes_before": {
            "type": "integer",
            "title": "Total Recording Bytes Before"
          },
          "total_recording_bytes_after": {
            "type": "integer",
            "title": "Total Recording Bytes After"
          },
          "deleted_meeting_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Deleted Meeting Ids"
          },
          "failed_meeting_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Failed Meeting Ids"
          },
          "skipped_meeting_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skipped Meeting Ids"
          }
        },
        "type": "object",
        "required": [
          "candidate_meeting_ids",
          "delete_count",
          "delete_recording_bytes",
          "total_recording_bytes_before",
          "total_recording_bytes_after",
          "deleted_meeting_ids",
          "failed_meeting_ids",
          "skipped_meeting_ids"
        ],
        "title": "RecordingCleanupExecuteResponse"
      },
      "RecordingCleanupPreviewResponse": {
        "properties": {
          "candidate_meeting_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Candidate Meeting Ids"
          },
          "delete_count": {
            "type": "integer",
            "title": "Delete Count"
          },
          "delete_recording_bytes": {
            "type": "integer",
            "title": "Delete Recording Bytes"
          },
          "total_recording_bytes_before": {
            "type": "integer",
            "title": "Total Recording Bytes Before"
          },
          "total_recording_bytes_after": {
            "type": "integer",
            "title": "Total Recording Bytes After"
          }
        },
        "type": "object",
        "required": [
          "candidate_meeting_ids",
          "delete_count",
          "delete_recording_bytes",
          "total_recording_bytes_before",
          "total_recording_bytes_after"
        ],
        "title": "RecordingCleanupPreviewResponse"
      },
      "RecordingCleanupRequest": {
        "properties": {
          "cutoff_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cutoff Date"
          },
          "max_total_bytes": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Total Bytes"
          }
        },
        "type": "object",
        "title": "RecordingCleanupRequest",
        "description": "A user-requested cleanup policy; omitted values disable that condition."
      },
...
      "UpdateMeetingTitleResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "UpdateMeetingTitleResponse"
      },
      "UpdateTitleRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "UpdateTitleRequest"
      },
...
    }
  }
}